+++ app/src/main/java/com/example/instagramrecovery/presentation/controller/DataActivityController.java
@@ -1,0 +1,8 @@
+package com.example.instagramrecovery.presentation.controller;
+
+public class DataActivityController {
+
+    public void onStart(){
+
+    }
+}
