+++ app/src/main/java/com/example/instagramrecovery/presentation/view/InstagramPost.java
@@ -19,7 +19,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
 
 import com.example.instagramrecovery.R;
-import com.example.instagramrecovery.presentation.model.SaveImageHelper;
+import com.example.instagramrecovery.SaveImageHelper;
 import com.squareup.picasso.Picasso;
 
 import java.util.UUID;
